initial support for 122mkii
