+++ Makefile
@@ -1,5 +1,5 @@
-obj-m += snd-usb-us144mkii.o
-snd-usb-us144mkii-y := us144mkii.o us144mkii_pcm.o us144mkii_playback.o us144mkii_capture.o us144mkii_midi.o us144mkii_controls.o
+obj-m += snd-usb-usx2y-mkii.o
+snd-usb-usx2y-mkii-y := usx2y_mkii.o usx2y_mkii_pcm.o usx2y_mkii_playback.o usx2y_mkii_capture.o usx2y_mkii_midi.o usx2y_mkii_controls.o
 
 all:
 	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
