+++ tascam_controls/resources/device_144mkii.png
